git-svn-id: file:///mnt/data/Workspace/zie87/game_development/tmp/hare-game-engine/p/trunk@269 67360d68-41fe-4630-88ca-a5c15c9c3197

+++ graphics/Animation.cpp
@@ -105,7 +105,7 @@ namespace hare
         frameMove();
 
 		if (rectPos.isEmpty()){
-            getCanvas()->drawImage((int)rectPos.minX, (int)rectPos.minY, animFrame->getMaterial());		
+            getCanvas()->drawImage(rectPos.minX, rectPos.minY, animFrame->getMaterial());		
 		}else{
 		    getCanvas()->drawImage(rectPos, animFrame->getMaterial());
 		}
